backend/requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
